unsigned solution

+++ h1/unsigned/unsigned.h
@@ -15,6 +15,6 @@ void ushortIntervalli(unsigned short sade, unsigned short askel);
 
 void uintIntervalli(unsigned int sade, unsigned int askel);
 
-void ulongIntervalli(unsigned long long sade, unsigned long long askel);
+void ulongIntervalli(unsigned long sade, unsigned long askel);
 
 #endif
